+++ backend/main.py
@@ -5,6 +5,11 @@ from pydantic import BaseModel
 from fastapi.middleware.cors import CORSMiddleware
 from datetime import datetime
 from service import get_current_user, generate_invite_code
+from receiptParsing.chatGPTParse import getChatGPTResponse
+
+from Join import app as join_router
+from Users import app as users_router
+from typing import List, Optional, Any
 
 # Initialize routers
 join_router = APIRouter()
@@ -175,6 +180,7 @@ async def send_fridge_invite(fridge_invite_dto: FridgeInviteDTO, current_user = 
         "message": "Invitation processing completed",
         "results": results
     }
+    
 
 # Accept fridge invite
 @join_router.post("/accept-invite/")
@@ -262,6 +268,7 @@ class FridgeCreate(BaseModel):
 
 @app.post("/fridges")
 def create_fridge(fridge: FridgeCreate):
+
     try:
         # Insert the fridge and get the response
         response = supabase.table("fridges").insert({
@@ -307,3 +314,15 @@ def get_fridges():
     except Exception as e:
         print(f"Error fetching fridges: {str(e)}")
         raise HTTPException(status_code=500, detail=str(e))
+    
+class Receipt(BaseModel):
+    base64Image: str
+
+@app.post("/parse-receipt")
+def parse_receipt(receipt: Receipt):
+    try:
+        return getChatGPTResponse(receipt.base64Image);
+    except Exception as e:
+        error_msg = f"Error parsing receipt: {str(e)}"
+        print(error_msg)
+        raise HTTPException(status_code=500, detail=error_msg)
